arduino-mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arduino-mcp-server"
version = "0.1.0"
description = "Natural language driven Arduino development MCP server"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "pyserial>=3.5",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests, no external dependencies",
    "integration: Integration tests, require arduino-cli",
    "hardware: Hardware tests, require physical board",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
arduino-mcp-server = "arduino_mcp_server.server:main"
